A possible fix for compiling on 32bit systems

+++ include/luxrays/luxrays.h
@@ -74,7 +74,7 @@ typedef unsigned char u_char;
 typedef unsigned short u_short;
 typedef unsigned int u_int;
 typedef unsigned long u_long;
-typedef unsigned long u_longlong;
+typedef unsigned long long u_longlong;
 
 /*!
  * \namespace luxrays
